Establish HTML, get images and create style page
The HTML page was established based on the guide image of what
the landing page should look like. The images that will be used already
got edited on GIMP to fit the content size and are located on the
designated folder. The CSS page was created, but not edited yet.